requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
redis==4.3.4
validators==0.20.0
flask-migrate==3.1.0
python_dotenv==0.20.0
flask-jwt-extended==4.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
marshmallow==3.17.1
flasgger==0.9.5
pydantic==1.10.2
flask-parameter-validation==2.1.0